2.7 Compat and strip Cook's D measures
This commit removes the matmul operator from Python 3 that you can't use in
Python 2.

In addition, I removed the computations of Cook's D, partially because they
weren't complete in the notebook and partially because, if done correctly for
nonlinear regression, it'd be longer to implement than to really be useful at
this moment.